requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
boto3==1.35.24
botocore==1.35.24
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
da-vinci==0.4.0
Django==5.0.6
django-browser-reload==1.12.1
django-environ==0.11.2
django-storages==1.14.4
django-tailwind==3.8.0
django-thumbnails==0.8.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
pillow==10.4.0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.0
rich==13.6.0
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
urllib3==2.2.1
